+++ node/src/types/kernel/api/notebook/getNotebookConf/index.d.ts
@@ -17,7 +17,7 @@
 
 
 /* directories */
-export * as getNotebookConf from "./getNotebookConf";
-export * as lsNotebooks from "./lsNotebooks";
 
 /* flles */
+export * from "./payload";
+export * from "./response";